TID-31/UnitTest - text mesh font rendering (#91)
* Bugfix-VertexBufferLayoutSetup

* Correctted vertex buffer layout setup for mesh_t and text_mesh_t

* Added legal notice on top of the source files

+++ include/renderer/internal/vulkan/vulkan_vertex_buffer_layout_description.h
@@ -328,8 +328,8 @@ typedef struct vulkan_vertex_buffer_layout_description_t
 	u32 input_rate;
 	u32 size;
 	u32 attribute_count;
-	u16 binding;
-	u16* attribute_locations;
+	u32 binding;
+	u32* attribute_locations;
 	VkFormat* attribute_formats;
 	u32* attribute_offsets;
 } vulkan_vertex_buffer_layout_description_t;
